requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
